style(conditions): restructure components folder
sdf
