+++ client/app/lib/components/conditions/specifiers.ts
@@ -7,7 +7,7 @@ import {
   ConditionPostData,
 } from 'types/course/conditions';
 import AchievementCondition from './conditions/AchievementCondition';
-import { AnyCondition } from './conditions/AnyCondition';
+import { AnyCondition } from './AnyCondition';
 import AssessmentCondition from './conditions/AssessmentCondition';
 import LevelCondition from './conditions/LevelCondition';
 import SurveyCondition from './conditions/SurveyCondition';
